+++ src/components/GoalsComponents/AddGoalOptions/AddGoalOptions.tsx
@@ -1,15 +1,13 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import { useRecoilValue, useSetRecoilState } from "recoil";
 
 import plus from "@assets/images/plus.svg";
 import publicGoals from "@assets/images/publicGoals.svg";
-import archiveGoalsIcon from "@assets/images/archiveGoals.svg";
 
 import Loader from "@src/common/Loader";
-import { GoalItem } from "@src/models/GoalItem";
 import { darkModeState, displayToast } from "@src/store";
 import { displayAddGoal, displaySuggestionsModal } from "@src/store/GoalsState";
-import { getGoal, getGoalsFromArchive } from "@src/api/GoalsAPI";
+import { getGoal } from "@src/api/GoalsAPI";
 import { getPublicGoals } from "@src/services/goal.service";
 
 import "./AddGoalOptions.scss";
